pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nict-bw"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy>=2.0",
    "python-dotenv>=1.0",
    "requests>=2.31"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["nictbw*"]
